+++ stories/demos/Themes.story.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { darkTheme, GraphCanvas } from '../../src';
 import { simpleEdges, simpleNodes } from '../assets/demo';
 
@@ -16,7 +17,7 @@ export const DarkTheme = () => (
   />
 );
 
-const CustomThemeStory = (args) => (
+const CustomThemeStory = args => (
   <GraphCanvas
     {...args}
     nodes={simpleNodes}
@@ -47,7 +48,7 @@ CustomTheme.args = {
       label: {
         stroke: '#1E2026',
         color: '#ACBAC7',
-        activeColor: '#1DE9AC',
+        activeColor: '#1DE9AC'
       }
     },
     menu: {